- Removing the synchronization across devices of reductions computed in various spots. Reductions ops used to be nondeterministic on the GPU in XLA, but apparently not anymore.
- Removing the pmap_axis_name argument in the multiply_* and _update_curvature_matrix_estimate methods in the curvature blocks since it's no longer used there. Also removing it from _update_cache methods since it's hasn't been used there for a while either.

- Minor internal refactor of optimizer, and of pi_adjusted_kronecker_inverse.

- Updating patches_moments_explicit to propagate dtypes properly in one spot. (It's unclear if this matters.)

- Adding 'per_device_stats_to_log' option to example experiment class which cause the listed statistics to be logged on a per-device basis (which is useful for checking that the devices remain properly synchronized).

PiperOrigin-RevId: 520611599

+++ kfac_jax/_src/patches_second_moment.py
@@ -653,17 +653,25 @@ def patches_moments_explicit(
 
   def loop_cond(args):
     return args[0] < c
+
   def loop_body(args):
+
     i, image, matrix_result, vector_result = args
+
     matrix_update, vector_update = general_loop_body(i, image)
+
     matrix_result = lax.dynamic_update_slice(
         matrix_result, matrix_update, (0, 0, 0, 0, i))
+
     vector_result = lax.dynamic_update_slice(
         vector_result, vector_update, (0, 0, i))
+
     return i + 1, image, matrix_result, vector_result
+
   init_vals = (0, inputs,
-               jnp.zeros(matrix_target_shape),
-               jnp.zeros(vector_target_shape))
+               jnp.zeros(matrix_target_shape, dtype=inputs.dtype),
+               jnp.zeros(vector_target_shape, dtype=inputs.dtype))
+
   return lax.while_loop(loop_cond, loop_body, init_vals)[-2:]
